requirements.txt
# Snake Venom Classifier

numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
opencv-python>=4.8.0
matplotlib>=3.7.0
joblib>=1.3.0
Pillow>=10.0.0

# Optional dependencies for advanced features
# torch>=2.0.0
# tensorflow>=2.13.0
streamlit>=1.25.0
